requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.5
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.1
inflection==0.5.1
jsonschema==3.2.0
psycopg2==2.8.6
PyJWT==2.0.1
pyrsistent==0.17.3
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
